+++ CHANGELOG.rst
@@ -7,7 +7,8 @@ CHANGELOG
 
 * Update development dependencies and better pytest practices for mocks/spies (`#54`_)
 * Rename ``QuickBase-`` class prefixes to ``Quickbase-`` and make aliases  (`#54`_)
-* Switch to Git Flow branching model with a main branch (`#54`_)
+* Switch to Git Flow branching model with a ``main`` branch (`#54`_)
+* Add more documentation for query endpoint usage (`#53`_)
 
 
 `0.4.0`_ (2021-6-30)
@@ -75,6 +76,7 @@ CHANGELOG
 .. _`#44`: https://github.com/tkutcher/quickbase-client/issues/44
 .. _`#45`: https://github.com/tkutcher/quickbase-client/issues/45
 .. _`#47`: https://github.com/tkutcher/quickbase-client/issues/47
+.. _`#53`: https://github.com/tkutcher/quickbase-client/issues/53
 .. _`#54`: https://github.com/tkutcher/quickbase-client/issues/54